requirements.txt
streamlit
torch
torchvision
Pillow
numpy
intel-openmp
msvc-runtime
